+++ src/methods/$includes_/async-iterable-includes_.mjs
@@ -6,7 +6,7 @@
  * More information can be found in CONTRIBUTING.md
  */
 
-import { asyncWindow } from '../$window/async-window';
+import { asyncLeadingWindow } from '../$leading-window/async-leading-window';
 import asyncToAnySubseq from '../../internal/async-to-any-subseq';
 import { iterableStartsWith_ } from '../$starts-with_/iterable-starts-with_';
 const startsWithConfig = {
@@ -18,7 +18,7 @@ export async function asyncIterableIncludes_(iterable, config, value) {
   const maxMatchLength = subseqs.reduce((max, { length }) => Math.max(max, length), 1);
   let hasItems = false;
 
-  for await (const buffer of asyncWindow(iterable, maxMatchLength)) {
+  for await (const buffer of asyncLeadingWindow(iterable, maxMatchLength)) {
     if (iterableStartsWith_(buffer, startsWithConfig, subseqs)) {
       return true;
     }