+++ src/methods/$includes_/iterable-includes_.mjs
@@ -6,7 +6,7 @@
  * More information can be found in CONTRIBUTING.md
  */
 
-import { window } from '../$window/window';
+import { leadingWindow } from '../$leading-window/leading-window';
 import toAnySubseq from '../../internal/to-any-subseq';
 import { iterableStartsWith_ } from '../$starts-with_/iterable-starts-with_';
 const startsWithConfig = {
@@ -18,7 +18,7 @@ export function iterableIncludes_(iterable, config, value) {
   const maxMatchLength = subseqs.reduce((max, { length }) => Math.max(max, length), 1);
   let hasItems = false;
 
-  for (const buffer of window(iterable, maxMatchLength)) {
+  for (const buffer of leadingWindow(iterable, maxMatchLength)) {
     if (iterableStartsWith_(buffer, startsWithConfig, subseqs)) {
       return true;
     }